requirements.txt
streamlit
sentence-transformers
faiss-cpu
torchvision
pillow
numpy